Hook open/close notice on the website to CO2 levels instead of a button
I set up a service on aleksei.hackeriet.no which updates a .json file based on CO2 level in the hackerspace. If over 1000 - someone is there, if below - probably no one is there. Rationale: false positives which happened with the button (I came a couple of times when it was said on the website there was someone in the hackerspace and there was no one). I'll leave it up to you to decide whether you want to use it in addition to, or in combination with, a button.